+++ colourettu/tests/__init__.py
@@ -1,0 +1,9 @@
+"""
+Test suite for Coloutettu.
+"""
+
+from __future__ import absolute_import
+
+import test_colour
+import test_palette
+import test_setup